Move DPoP header creation to oidc-client-dpop-browser
The DPoP header creation function are now exported from the oidc module, and consumed by the browser one. That makes most of the DPoP-related code in the browser module useless: it will be removed in a next commit.
