+++ packages/browser/src/dpop/DpopHeaderCreator.ts
@@ -32,12 +32,7 @@ import { inject, injectable } from "tsyringe";
 import IJoseUtility from "../jose/IJoseUtility";
 import { IDpopClientKeyManager } from "./DpopClientKeyManager";
 import { IUuidGenerator } from "../util/UuidGenerator";
-import {
-  /*signJWT,*/ privateJWKToPublicJWK,
-} from "../jose/IsomorphicJoseUtility";
-import { signJWT } from "@inrupt/oidc-dpop-client-browser";
-import { v4 } from "uuid";
-import { JSONWebKey } from "jose";
+import { createHeaderToken } from "@inrupt/oidc-dpop-client-browser";
 
 export interface IDpopHeaderCreator {
   /**
@@ -46,47 +41,6 @@ export interface IDpopHeaderCreator {
    * @param method The HTTP method that is being used
    */
   createHeaderToken(audience: URL, method: string): Promise<string>;
-}
-
-/**
- * Normalizes a URL in order to generate the DPoP token based on a consistent scheme.
- * @param audience The URL to normalize.
- * @returns The normalized URL as a string.
- * @hidden
- */
-export function normalizeHtu(audience: URL): string {
-  return `${audience.origin}${audience.pathname}`;
-}
-
-/**
- * Creates a DPoP header according to https://tools.ietf.org/html/draft-fett-oauth-dpop-04,
- * based on the target URL and method, using the provided key.
- * @param audience Target URL.
- * @param method HTTP method allowed.
- * @param key Key used to sign the token.
- * @returns A JWT that can be used as a DPoP Authorization header.
- */
-export async function createHeaderToken(
-  audience: URL,
-  method: string,
-  key: JSONWebKey
-): Promise<string> {
-  return signJWT(
-    {
-      htu: normalizeHtu(audience),
-      htm: method,
-      jti: v4(),
-    },
-    key,
-    {
-      header: {
-        jwk: privateJWKToPublicJWK(key),
-        typ: "dpop+jwt",
-      },
-      expiresIn: "1 hour",
-      algorithm: "ES256",
-    }
-  );
 }
 
 /**
@@ -100,8 +54,6 @@ export default class DpopHeaderCreator implements IDpopHeaderCreator {
     private dpopClientKeyManager: IDpopClientKeyManager,
     @inject("uuidGenerator") private uuidGenerator: IUuidGenerator
   ) {}
-
-  public normalizeHtu = normalizeHtu;
 
   async createHeaderToken(audience: URL, method: string): Promise<string> {
     // TODO: update for multiple signing abilities
